+++ Samples/iOS-Swift/iOS-Swift/Tools/iOS-Swift-Bridging-Header.h
@@ -2,4 +2,3 @@
 #import "SentryBenchmarking.h"
 #import "SentryExposure.h"
 #import <Sentry/PrivateSentrySDKOnly.h>
-#import <Sentry/SentryDsn+Private.h>