+++ Tests/SentryTests/Integrations/MetricKit/SentryMetricKitIntegrationTests.swift
@@ -17,7 +17,6 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     var callStackTreePerThread: SentryMXCallStackTree!
     var callStackTreeNotPerThread: SentryMXCallStackTree!
     var timeStampBegin: Date!
-    var timeStampEnd: Date!
     
     override func setUpWithError() throws {
         try super.setUpWithError()
@@ -28,10 +27,7 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
         let contentsNotPerThread = try contentsOfResource("MetricKitCallstacks/not-per-thread")
         callStackTreeNotPerThread = try SentryMXCallStackTree.from(data: contentsNotPerThread)
         
-        // Starting from iOS 15 MetricKit payloads are delivered immediately, so
-        // timeStamp and timeStampEnd match.
         timeStampBegin = SentryDependencyContainer.sharedInstance().dateProvider.date().addingTimeInterval(21.23)
-        timeStampEnd = timeStampBegin
     }
     
     override func tearDown() {
@@ -40,69 +36,76 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     }
 
     func testOptionEnabled_MetricKitManagerInitialized() {
-            let sut = SentryMetricKitIntegration()
-            
-            givenInstalledWithEnabled(sut)
-            
-            XCTAssertNotNil(Dynamic(sut).metricKitManager as SentryMXManager?)
+          let options = Options()
+          options.enableMetricKit = true
+          let sut = SentryMetricKitIntegration(with: options, dependencies: ())
+          XCTAssertNotNil(sut)
     }
     
     func testOptionDisabled_MetricKitManagerNotInitialized() {
-            let sut = SentryMetricKitIntegration()
-            
-            sut.install(with: Options())
-            
-            XCTAssertNil(Dynamic(sut).metricKitManager as SentryMXManager?)
-    }
-    
-    func testUninstall_MetricKitManagerSetToNil() {
-            let sut = SentryMetricKitIntegration()
-            
-            let options = Options()
-            options.enableMetricKit = true
-            sut.install(with: options)
-            sut.uninstall()
-            
-            XCTAssertNil(Dynamic(sut).metricKitManager as SentryMXManager?)
+          let options = Options()
+          let sut = SentryMetricKitIntegration(with: options, dependencies: ())
+          XCTAssertNil(sut)
     }
     
     func testMXCrashPayloadReceived() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut)
+        let options = Options()
+        options.enableMetricKit = true
+        let sut = SentryMXManager(inAppLogic: SentryInAppLogic(inAppIncludes: []), attachDiagnosticAsAttachment: false, disableCrashDiagnostics: false)
+
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/per-thread")
+        let crashDiagnostic = TestMXCrashDiagnostic()
+        crashDiagnostic.overrides.callStackTree = callStackTree
+        payload.overrides.crashDiagnostics = [crashDiagnostic]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.didReceive([payload])
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            mxDelegate.didReceiveCrashDiagnostic(MXCrashDiagnostic(), callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
-            
-            try assertPerThread(exceptionType: "MXCrashDiagnostic", exceptionValue: "MachException Type:(null) Code:(null) Signal:(null)", exceptionMechanism: "MXCrashDiagnostic", handled: false)
+        try assertPerThread(exceptionType: "MXCrashDiagnostic", exceptionValue: "MachException Type:nil Code:nil Signal:nil", exceptionMechanism: "MXCrashDiagnostic", handled: false)
     }
     
     func testAttachDiagnosticAsAttachment() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut) { $0.enableMetricKitRawPayload = true }
+        let options = Options()
+        options.enableMetricKit = true
+        options.enableMetricKitRawPayload = true
+        let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            let diagnostic = MXCrashDiagnostic()
-            mxDelegate.didReceiveCrashDiagnostic(diagnostic, callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/per-thread")
+        let hangDiagnostic = TestMXHangDiagnostic()
+        hangDiagnostic.overrides.callStackTree = callStackTree
+        payload.overrides.hangDiagnostic = [hangDiagnostic]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             try assertEventWithScopeCaptured { _, scope, _ in
                 let diagnosticAttachment = scope?.attachments.first { $0.filename == "MXDiagnosticPayload.json" }
                 
-                XCTAssertEqual(diagnosticAttachment?.data, diagnostic.jsonRepresentation())
+                XCTAssertEqual(diagnosticAttachment?.data, hangDiagnostic.jsonRepresentation())
             }
     }
     
     func testDontAttachDiagnosticAsAttachment() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
+        let options = Options()
+        options.enableMetricKit = true
+        let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            let diagnostic = MXCrashDiagnostic()
-            mxDelegate.didReceiveCrashDiagnostic(diagnostic, callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/per-thread")
+        let hangDiagnostic = TestMXHangDiagnostic()
+        hangDiagnostic.overrides.callStackTree = callStackTree
+        payload.overrides.hangDiagnostic = [hangDiagnostic]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             try assertEventWithScopeCaptured { _, scope, _ in
                 let diagnosticAttachment = scope?.attachments.first { $0.filename == "MXDiagnosticPayload.json" }
@@ -114,13 +117,19 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     func testSetInAppIncludes_AppliesInAppToStackTrace() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut) { options in
-                options.add(inAppInclude: "iOS-Swift")
-            }
+        let options = Options()
+        options.enableMetricKit = true
+        options.add(inAppInclude: "iOS-Swift")
+        let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            mxDelegate.didReceiveCrashDiagnostic(MXCrashDiagnostic(), callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/per-thread")
+        let hangDiagnostic = TestMXHangDiagnostic()
+        hangDiagnostic.overrides.callStackTree = callStackTree
+        payload.overrides.hangDiagnostic = [hangDiagnostic]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             try assertEventWithScopeCaptured { event, _, _ in
                 let stacktrace = try XCTUnwrap( event?.threads?.first?.stacktrace)
@@ -131,26 +140,21 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             }
     }
     
-    func testCPUExceptionDiagnostic_PerThread() throws {
-            givenSDKWithHubWithScope()
-            
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut)
-            
-            let mxDelegate = try XCTUnwrap(sut as? SentryMXManagerDelegate)
-            mxDelegate.didReceiveCpuExceptionDiagnostic(TestMXCPUExceptionDiagnostic(), callStackTree: callStackTreePerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
-            
-            assertNothingCaptured()
-    }
-    
     func testCPUExceptionDiagnostic_NotPerThread() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut)
+            let options = Options()
+            options.enableMetricKit = true
+            let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            mxDelegate.didReceiveCpuExceptionDiagnostic(TestMXCPUExceptionDiagnostic(), callStackTree: callStackTreeNotPerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/not-per-thread")
+        let cpuException = TestMXCPUExceptionDiagnostic()
+        cpuException.overrides.callStackTree = callStackTree
+        payload.overrides.cpuDiagnostic = [cpuException]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             try assertNotPerThread(exceptionType: "MXCPUException", exceptionValue: "MXCPUException totalCPUTime:2.2 ms totalSampledTime:5.5 ms", exceptionMechanism: "mx_cpu_exception")
     }
@@ -158,14 +162,18 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     func testCPUExceptionDiagnostic_OnlyOneFrame() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut)
+            let options = Options()
+            options.enableMetricKit = true
+            let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let contents = try contentsOfResource("MetricKitCallstacks/not-per-thread-only-one-frame")
-            let callStackTree = try SentryMXCallStackTree.from(data: contents)
-            
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            mxDelegate.didReceiveCpuExceptionDiagnostic(TestMXCPUExceptionDiagnostic(), callStackTree: callStackTree, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/not-per-thread-only-one-frame")
+        let cpuException = TestMXCPUExceptionDiagnostic()
+        cpuException.overrides.callStackTree = callStackTree
+        payload.overrides.cpuDiagnostic = [cpuException]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             guard let client = SentrySDKInternal.currentHub().getClient() as? TestClient else {
                 XCTFail("Hub Client is not a `TestClient`")
@@ -173,10 +181,9 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             }
             
             let invocations = client.captureEventWithScopeInvocations.invocations
-            XCTAssertEqual(2, client.captureEventWithScopeInvocations.count)
+            XCTAssertEqual(1, client.captureEventWithScopeInvocations.count)
             
             try assertEvent(event: try XCTUnwrap(invocations.first).event)
-            try assertEvent(event: try XCTUnwrap(invocations.element(at: 1)).event)
             
             func assertEvent(event: Event) throws {
                 let sentryFrames = try XCTUnwrap(event.threads?.first?.stacktrace?.frames, "Event has no frames.")
@@ -192,11 +199,18 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     func testDiskWriteExceptionDiagnostic() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut)
+            let options = Options()
+            options.enableMetricKit = true
+            let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            mxDelegate.didReceiveDiskWriteExceptionDiagnostic(TestMXDiskWriteExceptionDiagnostic(), callStackTree: callStackTreeNotPerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/not-per-thread-only-one-frame")
+        let diskWriteException = TestMXDiskWriteExceptionDiagnostic()
+        diskWriteException.overrides.callStackTree = callStackTree
+        payload.overrides.diskWriteDiagnostic = [diskWriteException]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             try assertNotPerThread(exceptionType: "MXDiskWriteException", exceptionValue: "MXDiskWriteException totalWritesCaused:5.5 Mib", exceptionMechanism: "mx_disk_write_exception")
     }
@@ -204,20 +218,20 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
     func testHangDiagnostic() throws {
             givenSDKWithHubWithScope()
             
-            let sut = SentryMetricKitIntegration()
-            givenInstalledWithEnabled(sut)
+            let options = Options()
+            options.enableMetricKit = true
+            let sut = try XCTUnwrap(SentryMetricKitIntegration(with: options, dependencies: ()))
             
-            let mxDelegate = sut as! SentryMXManagerDelegate
-            mxDelegate.didReceiveHangDiagnostic(TestMXHangDiagnostic(), callStackTree: callStackTreeNotPerThread, timeStampBegin: timeStampBegin, timeStampEnd: timeStampEnd)
+        let payload = TestMXDiagnosticPayload()
+        let callStackTree = TestMXCallStackTree()
+        callStackTree.overrides.jsonRepresentation = try contentsOfResource("MetricKitCallstacks/not-per-thread-only-one-frame")
+        let hangDiagnostic = TestMXHangDiagnostic()
+        hangDiagnostic.overrides.callStackTree = callStackTree
+        payload.overrides.hangDiagnostic = [hangDiagnostic]
+        payload.overrides.timeStampBegin = timeStampBegin
+        sut.mxManager.didReceive([payload])
             
             try assertNotPerThread(exceptionType: "MXHangDiagnostic", exceptionValue: "MXHangDiagnostic hangDuration:6.6 sec", exceptionMechanism: "mx_hang_diagnostic")
-    }
-    
-    private func givenInstalledWithEnabled(_ integration: SentryMetricKitIntegration, optionsBlock: (Options) -> Void = { _ in }) {
-        let options = Options()
-        options.enableMetricKit = true
-        optionsBlock(options)
-        integration.install(with: options)
     }
     
     private func givenSDKWithHubWithScope() {
@@ -235,12 +249,7 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
             XCTAssertEqual(callStackTreePerThread.callStacks.count, event?.threads?.count)
             XCTAssertEqual(timeStampBegin, event?.timestamp)
             
-            for callSack in callStackTreePerThread.callStacks {
-                var flattenedRootFrames = callSack.flattenedRootFrames
-                flattenedRootFrames.reverse()
-                
-                try assertFrames(frames: flattenedRootFrames, event: event, exceptionType, exceptionValue, exceptionMechanism, handled: handled)
-            }
+            try assertFrames(event: event, exceptionType, exceptionValue, exceptionMechanism, framesCount: 3, handled: handled)
         }
     }
     
@@ -260,65 +269,16 @@ final class SentryMetricKitIntegrationTests: SentrySDKIntegrationTestsBase {
         }
         
         let invocations = client.captureEventWithScopeInvocations.invocations
-        XCTAssertEqual(4, client.captureEventWithScopeInvocations.count, "Client expected to capture 2 events.")
-        
-        let firstEvent = try XCTUnwrap(invocations.first).event
-        let secondEvent = try XCTUnwrap(invocations.element(at: 1)).event
-        let thirdEvent = try XCTUnwrap(invocations.element(at: 2)).event
-        let fourthEvent = try XCTUnwrap(invocations.element(at: 3)).event
-        
-        for event in invocations.map({ $0.event }) {
-            XCTAssertEqual(timeStampBegin, event.timestamp)
-            XCTAssertEqual(false, try XCTUnwrap(event.threads?.first).crashed)
-        }
-        
-        let allFrames = try XCTUnwrap(callStackTreeNotPerThread.callStacks.first?.flattenedRootFrames, "CallStackTree has no call stack.")
-        
-        // Overview of stacktrace
-        // | frame 0 |
-        //      | frame 1 |
-        //          | frame 2 |
-        //          | frame 3 |
-        //              | frame 4 |
-        //              | frame 5 |
-        //              | frame 6 |     -> stack trace consists of [0,1,3,4,5,6]
-        //          | frame 7 |
-        //          | frame 8 |         -> stack trace consists of [0,1,2,3,7,8]
-        //      | frame 9 |             -> stack trace consists of [0,1,9]
-        // | frame 10 |
-        //      | frame 11 |
-        //          | frame 12 |
-        //          | frame 13 |    -> stack trace consists of [10,11,12,13]
-        
-        let firstEventFrames = [0, 1, 2, 3, 4, 5, 6].map { allFrames[$0] }
-        let secondEventFrames = [0, 1, 2, 3, 7, 8].map { allFrames[$0] }
-        let thirdEventFrames = [0, 1, 9].map { allFrames[$0] }
-        let fourthEventFrames = [10, 11, 12, 13].map { allFrames[$0] }
-        
-        try assertFrames(frames: firstEventFrames, event: firstEvent, exceptionType, exceptionValue, exceptionMechanism, debugMetaCount: 3)
-        try assertFrames(frames: secondEventFrames, event: secondEvent, exceptionType, exceptionValue, exceptionMechanism, debugMetaCount: 3)
-        try assertFrames(frames: thirdEventFrames, event: thirdEvent, exceptionType, exceptionValue, exceptionMechanism, debugMetaCount: 3)
-        try assertFrames(frames: fourthEventFrames, event: fourthEvent, exceptionType, exceptionValue, exceptionMechanism, debugMetaCount: 3)
+        XCTAssertEqual(1, invocations.count, "Client expected to capture 1 event.")
     }
     
-    private func assertFrames(frames: [SentryMXFrame], event: Event?, _ exceptionType: String, _ exceptionValue: String, _ exceptionMechanism: String, handled: Bool = true, debugMetaCount: Int = 2) throws {
+    private func assertFrames(event: Event?, _ exceptionType: String, _ exceptionValue: String, _ exceptionMechanism: String, framesCount: Int, handled: Bool = true, debugMetaCount: Int = 2) throws {
         let sentryFrames = try XCTUnwrap(event?.threads?.first?.stacktrace?.frames, "Event has no frames.")
-        XCTAssertEqual(frames.count, sentryFrames.count)
+        XCTAssertEqual(framesCount, sentryFrames.count)
         
         XCTAssertEqual(1, event?.exceptions?.count)
         let exception = try XCTUnwrap(event?.exceptions?.first, "Event has exception.")
-        XCTAssertEqual(frames.count, exception.stacktrace?.frames.count)
-        
-        let exceptionFrames = try XCTUnwrap(exception.stacktrace?.frames, "Exception has no frames.")
-    
-        for i in 0..<frames.count {
-            let mxFrame = frames[i]
-            let sentryFrame = sentryFrames[i]
-            let sentryExceptionFrame = exceptionFrames[i]
-            assertFrame(mxFrame: mxFrame, sentryFrame: sentryFrame)
-            assertFrame(mxFrame: mxFrame, sentryFrame: sentryExceptionFrame)
-        }
-        
+
         XCTAssertEqual(exceptionType, exception.type)
         XCTAssertEqual(exceptionValue, exception.value)
         XCTAssertEqual(exceptionMechanism, exception.mechanism?.type)